feat: 리펙토링
Text -> CustomSplitter, Values 분리
